Renamed response types to token types
