+++ src/TokenTypes/BearerTokenType.php
@@ -9,11 +9,11 @@
  * @link        https://github.com/thephpleague/oauth2-server
  */
 
-namespace League\OAuth2\Server\ResponseTypes;
+namespace League\OAuth2\Server\TokenTypes;
 
 use Symfony\Component\HttpFoundation\Request;
 
-class BearerTokenResponseType extends AbstractResponseType
+class BearerTokenType extends AbstractTokenType
 {
     /**
      * {@inheritdoc}
